+++ R/qdrg.R
@@ -69,22 +69,21 @@
 #' 
 #' @export
 #' @examples
-#' \dontrun{
-#'   if(require(lme4)) {
-#'     # load a stored example having a posterior sample...
-#'     load(system.file("extdata", "cbpp.RData", package = "emmeans"))
-#'     
-#'     brg <- qdrg(cbind(incidence, size-incidence) ~ size + period, data = lme4::cbpp,
-#'         mcmc = cbpp.list$post.beta, link = "logit")
-#'     summary(brg, type = "response")
-#'   }
-#'   if (require(biglm)) {
-#'     bigmod <- biglm(log(conc) ~ source + factor(percent), data = pigs)
-#'     
-#'     rg <- qdrg(object = bigmod, data = pigs)
-#'     summary(emmeans(rg, "source"), type = "response")
-#'   }
+#' if(require(coda) && require(lme4)) {
+#'   # Use a stored example having a posterior sample
+#'   # Model is based on the data in lme4::cbpp
+#'   
+#'   post <- readRDS(system.file("extdata", "cbpplist", package = "emmeans"))$post.beta
+#'   rg1 <- qdrg(~ size + period, data = lme4::cbpp, mcmc = post, link = "logit")
+#'   summary(rg1, type = "response")
 #' }
+#' if (require(biglm)) {
+#'   bigmod <- biglm(log(conc) ~ source + factor(percent), data = pigs)
+#'    
+#'   rg2 <- qdrg(object = bigmod, data = pigs)
+#'   summary(emmeans(rg2, "source"), type = "response")
+#' }
+#'
 qdrg = function(formula, data, coef, mcmc, vcov, object,
                 df, subset, weights, contrasts, link, qr, ...) {
     result = match.call(expand.dots = FALSE)