+++ R/emtrends.R
@@ -96,12 +96,12 @@ emtrends = function(model, specs, var, delta.var=.01*rng, data,
     estName = paste(var, "trend", sep=".") # Do now as I may replace var later
     
     if (missing(data)) {
-        data = try(.chk_recover_data (model, data = NULL))
+        data = try(recover_data (model, data = NULL))
         if (inherits(data, "try-error"))
             stop("Possible remedy: Supply the data used in the 'data' argument")
     }
     else # attach needed attributes to given data
-        data = .chk_recover_data(model, data = data)
+        data = recover_data(model, data = data)
     
     x = data[[var]]
     fcn = NULL   # differential
@@ -128,7 +128,7 @@ emtrends = function(model, specs, var, delta.var=.01*rng, data,
     }
     grid[[var]] = grid[[var]] + delta.var
     
-    basis = .chk_emm_basis(model, attr(data, "terms"), RG@model.info$xlev, grid, ...)
+    basis = emm_basis(model, attr(data, "terms"), RG@model.info$xlev, grid, ...)
     if (is.null(fcn))
         newlf = (basis$X - RG@linfct) / delta.var
     else {