+++ NEWS.md
@@ -4,6 +4,8 @@ emmeans 1.2.4.x
   * Courtesy exports `.recover_data()` and `.emm_basis()` to provide
     access for extension developers to all available methods
   * Streamlining of a stored example in `inst/extdata`
+  * Fix to `.all.vars()` that could cause errors when response variable
+    has a function call with character constants.
 
 
 emmeans 1.2.4